+++ UI/Sources/ShoppingCart/ShoppingCartTableViewController.swift
@@ -101,13 +101,6 @@ final class ShoppingCartTableViewController: UITableViewController {
         super.viewDidAppear(animated)
 
         self.shoppingCartDelegate?.track(.viewShoppingCart)
-
-        if !self.isBeingPresented && !self.isMovingToParent {
-            // whatever was covering us has been dismissed or popped
-
-            // re-send our current cart to the backend so that the supervisor can see us shopping again
-            CartEvent.cart(self.shoppingCart)
-        }
     }
 
     // MARK: notification handlers